Restore C++ locale when LocaleGuard goes out of scope
Only the C locale was being restored. I'm not sure if this was intentional but
I'm guessing not as it means that C++ level API like string_compose,
stringstreams etc can't be used for formatting/producing localized numeric data
as they will always use the global C++ locale/LC_NUMERIC setting which is "C"
after the first LocaleGuard is created. If it was intentional it should have
been set in main() and not as a side effect of using a LocaleGuard.

All instances where string_compose/stringstreams/etc have been used to format
numeric data with the assumption that the global C++ locale is always set to
"C" should now be using the pbd/string_convert.h API. Any instances where
string_compose/stringstream/etc was actually being used to generate a localized
message that contained numeric data should now be correctly producing localized
numeric formatting.

As a side note using std::ios::imbue doesn't work with all the
compilers/options/platforms we currently support from my testing and so is not
an option at this stage. It also seems more complicated that just setting and
leaving the global C and C++ locale as the users preferred locale and then just
using a specific API when we want numeric formatting that is portable which we
now have in pbd/string_convert.h

The only case where a LocaleGuard may be appropriate is for external code like
plugins where for whatever reason(broken/bad design/etc?) there is an
assumption or requirement that the global locale is "C".